move ResizeProfilingBenchmarks, use the ***ProfilingBenchmarks naming convention everywhere
